requirements.txt
absl-py==2.1.0
accelerate==0.34.2
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.4.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cython==3.0.11
debugpy==1.8.12
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
gdown==5.1.0
geotorch==0.3.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.4
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.1
importlib-metadata==8.2.0
importlib-resources==6.4.0
jinja2==3.1.4
jpeg4py==0.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kaggle==1.6.17
kiwisolver==1.4.5
lmdb==1.5.1
loguru==0.7.2
markdown==3.6
markupsafe==2.1.5
matplotlib==3.7.5
mkl-service==2.4.0
numpy==1.24.3
opencv-python==4.10.0.84
pandas==2.0.3
pillow==9.3.0
protobuf==5.27.3
psutil==6.0.0
pycocotools==2.0.7
pycparser==2.22
pyyaml==6.0.1
torchdiffeq==0.2.4
tqdm==4.66.4
transformers==4.44.2
triton==3.0.0
typing-extensions==4.11.0
wandb==0.17.5
webcolors==24.6.0
werkzeug==3.0.3
wget==3.2
wheel==0.43.0
visdom==0.2.4
yacs==0.1.8
safetensors==0.4.4
sympy==1.13.1
tensorboard==2.14.0
tensorboardx==2.6.2.2
